wa-sim/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
pyyaml>=6.0
rich>=13.0.0
python-dotenv>=1.0.0
